Task #55 - tryWith should work the same as Java's try-with-resources and remove TryWith - Changed: TryWith.SideEffect => SideEffect (TryWith has been removed)
